eqpot/wsgi.py
"""
WSGI config for eqpot project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/2.1/howto/deployment/wsgi/
"""

import os
from dotenv import load_dotenv


from django.core.wsgi import get_wsgi_application

BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
load_dotenv(os.path.join(project_folder, '.env'))


os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'eqpot.settings')

application = get_wsgi_application()
